Automatic-Exam-Trust-Score/requirements.txt
notebook
librosa
pandas
seaborn
numpy
tqdm
matplotlib
scikit-learn
tensorflow
django
resampy
python-magic-bin